Nectar - move communities, identity and messages interfaces to separate files
