+++ packages/nectar/src/sagas/communities/communities.slice.ts
@@ -1,9 +1,11 @@
 import { createSlice, EntityState, PayloadAction } from '@reduxjs/toolkit'
 import { StoreKeys } from '../store.keys'
 import { communitiesAdapter } from './communities.adapter'
-import { createRootCA } from '@zbayapp/identity'
-import { AsyncReturnType } from '../../utils/types/AsyncReturnType.interface'
-import { HiddenService, Identity, PeerId } from '../identity/identity.types'
+import {
+  ResponseCreateCommunityPayload,
+  ResponseRegistrarPayload,
+  StorePeerListPayload
+} from './communities.types'
 
 export class CommunitiesState {
   public currentCommunity: string = ''
@@ -29,55 +31,6 @@ export interface Community {
   onionAddress: string
   privateKey: string
   port: number
-}
-
-export interface Certificates {
-  certificate: string
-  key: string
-  CA: string[]
-}
-
-export interface InitCommunityPayload {
-  id: string
-  peerId: PeerId
-  hiddenService: HiddenService
-  certs: Certificates
-  peers?: string[]
-}
-
-export interface AddNewCommunityPayload {
-  id: string
-  name: string
-  CA: AsyncReturnType<typeof createRootCA> | {}
-  registrarUrl: string
-}
-
-export interface LaunchRegistrarPayload {
-  id: string
-  peerId: string
-  rootCertString: string
-  rootKeyString: string
-  privateKey: string
-  port?: number
-}
-
-export interface ResponseRegistrarPayload {
-  id: string
-  payload: Partial<Community>
-}
-
-export interface StorePeerListPayload {
-  communityId: string
-  peerList: string[]
-}
-
-export interface ResponseCreateCommunityPayload {
-  id: string
-  payload: Partial<Identity>
-}
-
-export interface ResponseLaunchCommunityPayload {
-  id: string
 }
 
 export const communitiesSlice = createSlice({
