+++ packages/nectar/src/sagas/communities/responseCreateCommunity/responseCreateCommunity.saga.ts
@@ -1,8 +1,8 @@
 import { put, call } from 'typed-redux-saga'
 import { identityActions } from '../../identity/identity.slice'
 import { generateDmKeyPair } from '../../../utils/cryptography/cryptography'
-import { ResponseCreateCommunityPayload } from '../../communities/communities.slice'
 import { PayloadAction } from '@reduxjs/toolkit'
+import { ResponseCreateCommunityPayload } from '../communities.types'
 
 export function* responseCreateCommunitySaga(
   action: PayloadAction<ResponseCreateCommunityPayload>
